moved RasterIntervalsLegend
