+++ src/components/controls/RasterIntervalsLegend.svelte
